Set become: yes
When doing lab 2 for MSES 602, I was getting repeated permissions issues. Making this change resolved everything for me. Not sure if it's just my instance itself, maybe I messed up a config setting, but I wanted to open a PR to show the diff in discussion, and maybe it'll resolve problems for other folks.